includes UserContext to only allow the user to delete their own comments

+++ src/components/CommentCard.jsx
@@ -1,15 +1,17 @@
 import { useContext } from "react"
-import {UserProvider } from "../contexts/LoggedInUser"
+import { UserContext } from "../contexts/LoggedInUser"
 import { deleteComment } from "../utils/api"
 
 const CommentCard = ({props}) => {
+    const { user } = useContext(UserContext)
+
 
     const deleteCommentButton = () => {
-        /** As it stands, this function simply invokes the function in the API,
-         * It could have some additional features, but it currently functions
-         *  albeit, in the most barebones implementation of the function.
-         */
-        deleteComment(props.comment_id)
+        if (props.author === user) {
+            deleteComment(props.comment_id)
+        } else {
+            console.log("That's not your comment to delete, pal!")
+        }
     }
 
 
@@ -17,7 +19,7 @@ const CommentCard = ({props}) => {
         
             <article className="comment-card">
                 <div className="details">
-                <button onClick={()=>{deleteCommentButton()}}>Delete Comment</button>
+                {props.author === user ? <button onClick={()=>{deleteCommentButton()}}>Delete Comment</button> : null}
                     <h3>{props.author}</h3>
                     <h3>{props.created_at}</h3>
                     <h3>{props.votes}</h3>
